pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vigil"
version = "1.0.0"
description = "A framework-agnostic, production-ready audit logging system for Python applications"
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
authors = [
    {name = "Your Name", email = "vigil@example.com"}
]
keywords = ["audit", "logging", "security", "compliance", "gdpr", "hipaa", "sox", "forensics"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Topic :: System :: Logging",
    "Topic :: Security",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
]
dependencies = [
    "pyyaml>=6.0",
    "python-dateutil>=2.8.0",
]

[project.scripts]
vigil-cli = "vigil.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
    "flake8>=6.0.0",
]
database = [
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "pymysql>=1.0.0",
]
cloud = [
    "boto3>=1.26.0",
    "azure-storage-blob>=12.0.0",
    "google-cloud-storage>=2.0.0",
]
queue = [
    "kafka-python>=2.0.0",
    "pika>=1.3.0",
    "redis>=4.5.0",
]
web = [
    "flask>=2.3.0",
    "fastapi>=0.100.0",
    "django>=4.2.0",
]
all = [
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "pymysql>=1.0.0",
    "boto3>=1.26.0",
    "azure-storage-blob>=12.0.0",
    "google-cloud-storage>=2.0.0",
    "kafka-python>=2.0.0",
    "pika>=1.3.0",
    "redis>=4.5.0",
    "flask>=2.3.0",
    "fastapi>=0.100.0",
    "django>=4.2.0",
]

[project.urls]
Homepage = "https://github.com/yourorg/vigil"
Documentation = "https://docs.example.com/vigil"
Repository = "https://github.com/yourorg/vigil"
"Bug Tracker" = "https://github.com/yourorg/vigil/issues"

[tool.setuptools]
packages = ["vigil"]

[tool.setuptools.package-data]
vigil = ["py.typed"]

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_optional = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers"
testpaths = [
    "tests",
]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"

[tool.coverage.run]
source = ["vigil"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__pycache__/*",
]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]
